feat: adjusted all pages styles, name and content
